refactor: move BrowserFragment related classes into same package
