+++ app/src/main/java/org/mozilla/focus/screenshot/CaptureRunnable.java
@@ -7,10 +7,10 @@ import android.view.View;
 import android.widget.Toast;
 
 import org.mozilla.focus.R;
-import org.mozilla.focus.fragment.BrowserFragment;
 import org.mozilla.focus.fragment.ScreenCaptureDialogFragment;
 import org.mozilla.focus.utils.AppConstants;
 import org.mozilla.focus.utils.Settings;
+import org.mozilla.rocket.browser.BrowserFragment;
 import org.mozilla.rocket.chrome.ChromeViewModel;
 
 import java.lang.ref.WeakReference;
