+++ app/src/main/res/layout/activity_main.xml
@@ -10,7 +10,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:ignore="MergeRootFrame">
-    <fragment android:name="org.mozilla.focus.fragment.BrowserFragment"
+    <fragment android:name="org.mozilla.rocket.browser.BrowserFragment"
         android:id="@+id/browser"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
